% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{birthweight}
\alias{birthweight}
\title{birthweight}
\usage{
data(birthweight)
}
\value{
A data frame with 20 rows and 4 variables.
}
\description{
birthweight
}
\examples{
data(birthweight)
}
\keyword{datasets}
