% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bmi2}
\alias{bmi2}
\title{bmi2}
\usage{
data(bmi2)
}
\value{
A data frame with 200 rows and 6 variables.
}
\description{
bmi2
}
\examples{
data(bmi)
}
\keyword{'Female'}}
\keyword{'Good',}
\keyword{'Healthy'}
\keyword{'Moderate'}
\keyword{'Poor'}}
\keyword{'Unhealthy'}}
\keyword{ID}
\keyword{\describe{}
\keyword{\item{age}{Subject's}
\keyword{\item{bmi}{Continuous}
\keyword{\item{diet}{Participants}
\keyword{\item{id}{Unique}
\keyword{\item{sex}{'Male'}
\keyword{\item{status}{Health}
\keyword{age}
\keyword{at}
\keyword{bmi}}
\keyword{capturing}
\keyword{datasets}
\keyword{diet:}
\keyword{experiment}}
\keyword{number}}
\keyword{of}
\keyword{or}
\keyword{participant}
\keyword{participants}
\keyword{status:}
\keyword{the}
\keyword{time}
\keyword{variable}
\keyword{}}
