% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{disease}
\alias{disease}
\title{disease}
\usage{
data(disease)
}
\value{
A data frame with 1000 rows and 7 variables.
}
\description{
disease
}
\examples{
data(disease)
}
\keyword{datasets}
