% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{us_teams}
\alias{us_teams}
\title{us_teams}
\usage{
data(us_teams)
}
\value{
A data frame with 50 rows and 3 variables.
}
\description{
us_teams
}
\examples{
data(us_teams)
}
\keyword{datasets}
