% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bmi2}
\alias{bmi2}
\title{bmi2}
\usage{
data(bmi2)
}
\value{
A data frame with 200 rows and 6 variables.
}
\description{
bmi2
}
\note{
The dataset contains the following 6 variables:
\describe{
\item{id}{Unique participant ID number}
\item{age}{Subject's age at time of experiment}
\item{bmi}{Continuous variable capturing the participants bmi}
\item{sex}{'Male' or 'Female'}
\item{diet}{Participants diet: 'Good', 'Moderate' or 'Poor'}
\item{status}{Health status: 'Healthy' or 'Unhealthy'}
}
}
\examples{
data(bmi2)
}
\keyword{datasets}
